+++ Lab_1/Task1/1.1/Dockerfile
@@ -2,6 +2,6 @@ FROM python
 
 WORKDIR "/home/egor/Documents/BSUIR/4sem_IGI_Python/Labs_Python_IGI/Lab_1/Task1/1.1"
 
-COPY ["PrintHello.py", "./"]
+COPY ["print_hello.py", "./"]
 
-CMD ["python", "PrintHello.py"]
+CMD ["python", "print_hello.py"]
