+++ Lab_1/Task1/1.1/print_hello.py
