+++ Lab_1/Task1/1.2/Dockerfile
@@ -2,8 +2,8 @@ FROM python
 
 WORKDIR "/home/egor/Documents/BSUIR/4sem_IGI_Python/Labs_Python_IGI/Lab_1/Task1/1.2"
 
-COPY ["/calculate.py", "./"]
-COPY ["Constant.py", "./"]
-COPY ["Util.py", "./"]
+COPY ["calculate.py", "./"]
+COPY ["constant.py", "./"]
+COPY ["util.py", "./"]
 
-ENTRYPOINT ["python", "calculate.py"]
+CMD ["python", "calculate.py"]
