+++ Lab_1/Task1/1.2/constant.py
