+++ Lab_1/Task1/1.3/Dockerfile
@@ -3,7 +3,7 @@ FROM python
 WORKDIR "/home/egor/Documents/BSUIR/4sem_IGI_Python/Labs_Python_IGI/Lab_1/Task1/1.3"
 
 COPY ["main.py", "./"]
-COPY ["Util.py", "./"]
+COPY ["util.py", "./"]
 
 
 CMD ["python", "main.py"]
